refactor(auth): RefreshSessionRepository 추출 (interface + DI Token)
P0-1 단계 1. AuthRepository(530줄, 15메서드)에서 refresh session
관련 5개 메서드를 분리하여 별도 Repository로 추출.

- IRefreshSessionRepository 인터페이스 + REFRESH_SESSION_REPOSITORY Symbol 토큰 정의
  (P1-1 신규 Repository는 interface + Token 패턴 적용)
- 5개 메서드 이전: createRefreshSession, findActiveRefreshSessionByHash,
  rotateRefreshSession, revokeRefreshSession, revokeAllRefreshSessions
- AuthService 7개 호출부 신규 repo 의존으로 전환
- 실DB 단위테스트 별도 파일(refresh-session.repository.spec.ts)로 분리
- Service 단위테스트 mock도 두 repo로 분리
