+++ src/features/auth/repositories/auth.repository.ts
@@ -264,120 +264,6 @@ export class AuthRepository {
   }
 
   /**
-   * refresh session을 생성한다(토큰은 hash만 저장).
-   *
-   * @param args 세션 생성 정보
-   */
-  async createRefreshSession(args: {
-    accountId: bigint;
-    tokenHash: string;
-    userAgent?: string;
-    ipAddress?: string;
-    expiresAt: Date;
-  }) {
-    return this.prisma.authRefreshSession.create({
-      data: {
-        account_id: args.accountId,
-        token_hash: args.tokenHash,
-        user_agent: args.userAgent ?? null,
-        ip_address: args.ipAddress ?? null,
-        expires_at: args.expiresAt,
-      },
-    });
-  }
-
-  /**
-   * refresh token hash로 세션을 조회한다(유효/미삭제만).
-   *
-   * @param tokenHash refresh token hash(sha256 hex)
-   */
-  async findActiveRefreshSessionByHash(tokenHash: string) {
-    const now = this.clock.now();
-    return this.prisma.authRefreshSession.findFirst({
-      where: {
-        token_hash: tokenHash,
-        revoked_at: null,
-        expires_at: { gt: now },
-      },
-    });
-  }
-
-  /**
-   * refresh session을 회전(rotation)한다.
-   *
-   * - 기존 세션 revoked
-   * - 신규 세션 생성
-   * - 기존 세션 replaced_by_session_id 연결
-   *
-   * @param args 회전 정보
-   */
-  async rotateRefreshSession(args: {
-    currentSessionId: bigint;
-    accountId: bigint;
-    newTokenHash: string;
-    userAgent?: string;
-    ipAddress?: string;
-    newExpiresAt: Date;
-  }) {
-    return this.prisma.$transaction(async (tx) => {
-      const now = this.clock.now();
-
-      const newSession = await tx.authRefreshSession.create({
-        data: {
-          account_id: args.accountId,
-          token_hash: args.newTokenHash,
-          user_agent: args.userAgent ?? null,
-          ip_address: args.ipAddress ?? null,
-          expires_at: args.newExpiresAt,
-        },
-      });
-
-      await tx.authRefreshSession.update({
-        where: { id: args.currentSessionId },
-        data: {
-          revoked_at: now,
-          replaced_by_session_id: newSession.id,
-          updated_at: now,
-        },
-      });
-
-      return newSession;
-    });
-  }
-
-  /**
-   * refresh session을 revoke 처리한다.
-   *
-   * @param sessionId session id
-   */
-  async revokeRefreshSession(sessionId: bigint) {
-    const now = this.clock.now();
-    return this.prisma.authRefreshSession.update({
-      where: { id: sessionId },
-      data: { revoked_at: now, updated_at: now },
-    });
-  }
-
-  /**
-   * 특정 계정의 활성 refresh session을 모두 revoke 처리한다.
-   *
-   * @param accountId account id
-   * @param now 기준 시각
-   */
-  async revokeAllRefreshSessions(accountId: bigint, now: Date): Promise<void> {
-    await this.prisma.authRefreshSession.updateMany({
-      where: {
-        account_id: accountId,
-        revoked_at: null,
-      },
-      data: {
-        revoked_at: now,
-        updated_at: now,
-      },
-    });
-  }
-
-  /**
    * access token 검증용으로 계정을 조회한다.
    *
    * @param accountId account id
